+++ R/splitEnrichment.R
@@ -114,19 +114,11 @@ splitEnrichment <- function(input.data,
   }
   
   col <- length(unique(enriched[,split.by]))
-  if (is.null(group.by)) {
-    plot <- ggplot(enriched, aes(x = ".", 
-                                 y = enriched[,gene.set], 
-                                 fill = enriched[,split.by])) 
-    check = 1
-  } else {
-    plot <- ggplot(enriched, aes(x = enriched[,group.by], 
+  plot <- ggplot(enriched, aes(x = enriched[,group.by], 
                                  y = enriched[,gene.set], 
                                  fill = enriched[,split.by])) + 
                   xlab(group.by) 
-    check = NULL
-    }
-  
+
   plot <- plot + 
           geom_split_violin(alpha=0.8, lwd= 0.25) +
           geom_boxplot(width=0.1, 
@@ -139,11 +131,6 @@ splitEnrichment <- function(input.data,
           scale_fill_manual(values = .colorizer(palette, col))+
           theme_classic() 
 
-    if (!is.null(check)) {
-      plot <- plot + 
-              theme(axis.title.x = element_blank(),
-                    axis.text.x = element_blank(),
-                    axis.ticks.x = element_blank())}
   if (!is.null(facet.by)) {
     plot <- plot + 
             facet_grid(as.formula(paste('. ~', facet.by)))